fix: move version/appname + directories to lapce-core
