+++ lapce-core/Cargo.toml
@@ -5,6 +5,7 @@ authors = ["Dongdong Zhou <dzhou121@gmail.com>"]
 edition = "2021"
 
 [dependencies]
+directories = "4.0.1"
 thiserror = "1.0"
 itertools = "0.10.3"
 log = "0.4.14"
