+++ lapce-core/src/directory.rs
@@ -2,13 +2,12 @@ use std::path::PathBuf;
 
 use directories::ProjectDirs;
 
-use crate::APPLICATION_NAME;
-
+use crate::meta::NAME;
 pub struct Directory {}
 
 impl Directory {
     fn project_dirs() -> Option<ProjectDirs> {
-        ProjectDirs::from("dev", "lapce", *APPLICATION_NAME)
+        ProjectDirs::from("dev", "lapce", *NAME)
     }
 
     // Get path of local data directory
@@ -22,13 +21,11 @@ impl Directory {
                 if !dir.exists() {
                     let _ = std::fs::create_dir_all(dir);
                 }
-
                 Some(dir.to_path_buf())
             }
             None => None,
         }
     }
-
     /// Get the path to logs directory
     /// Each log file is for individual application startup
     pub fn logs_directory() -> Option<PathBuf> {
@@ -37,13 +34,11 @@ impl Directory {
             if !dir.exists() {
                 let _ = std::fs::create_dir(&dir);
             }
-
             Some(dir)
         } else {
             None
         }
     }
-
     /// Directory to store proxy executables used on local
     /// host as well, as ones uploaded to remote host when
     /// connecting
@@ -53,13 +48,11 @@ impl Directory {
             if !dir.exists() {
                 let _ = std::fs::create_dir(&dir);
             }
-
             Some(dir)
         } else {
             None
         }
     }
-
     /// Get the path to the themes folder
     /// Themes are stored within as individual toml files
     pub fn themes_directory() -> Option<PathBuf> {
@@ -68,13 +61,11 @@ impl Directory {
             if !dir.exists() {
                 let _ = std::fs::create_dir(&dir);
             }
-
             Some(dir)
         } else {
             None
         }
     }
-
     // Get the path to plugins directory
     // Each plugin has own directory that contains
     // metadata file and plugin wasm
@@ -84,13 +75,11 @@ impl Directory {
             if !dir.exists() {
                 let _ = std::fs::create_dir(&dir);
             }
-
             Some(dir)
         } else {
             None
         }
     }
-
     // Config directory contain only configuration files
     pub fn config_directory() -> Option<PathBuf> {
         match Self::project_dirs() {
@@ -99,24 +88,20 @@ impl Directory {
                 if !dir.exists() {
                     let _ = std::fs::create_dir_all(dir);
                 }
-
                 Some(dir.to_path_buf())
             }
             None => None,
         }
     }
-
     pub fn local_socket() -> Option<PathBuf> {
         Self::data_local_directory().map(|dir| dir.join("local.sock"))
     }
-
     pub fn updates_directory() -> Option<PathBuf> {
         if let Some(dir) = Self::data_local_directory() {
             let dir = dir.join("updates");
             if !dir.exists() {
                 let _ = std::fs::create_dir(&dir);
             }
-
             Some(dir)
         } else {
             None
