+++ lapce-core/src/lib.rs
@@ -2,11 +2,13 @@ pub mod buffer;
 pub mod chars;
 pub mod command;
 pub mod cursor;
+pub mod directory;
 pub mod editor;
 pub mod encoding;
 pub mod indent;
 pub mod language;
 pub mod lens;
+pub mod meta;
 pub mod mode;
 pub mod movement;
 pub mod register;
