+++ lapce-data/src/data.rs
@@ -22,13 +22,14 @@ use itertools::Itertools;
 use lapce_core::{
     command::{FocusCommand, MultiSelectionCommand},
     cursor::{Cursor, CursorMode},
+    directory::Directory,
     editor::EditType,
+    meta,
     mode::MotionMode,
     movement::Movement,
     register::Register,
     selection::Selection,
 };
-use lapce_proxy::{directory::Directory, VERSION};
 use lapce_rpc::{
     buffer::BufferId,
     core::{CoreMessage, CoreNotification},
@@ -1173,7 +1174,7 @@ impl LapceTabData {
         match command {
             LapceWorkbenchCommand::RestartToUpdate => {
                 if let Some(release) = (*self.latest_release).clone() {
-                    if release.version != *VERSION {
+                    if release.version != *meta::VERSION {
                         if let Ok(process_path) = env::current_exe() {
                             ctx.submit_command(Command::new(
                                 LAPCE_UI_COMMAND,
