+++ lapce-data/src/db.rs
@@ -9,7 +9,7 @@ use std::{
 use anyhow::{anyhow, Result};
 use crossbeam_channel::{unbounded, Sender};
 use druid::{ExtEventSink, Point, Rect, Size, Vec2, WidgetId};
-use lapce_proxy::directory::Directory;
+use lapce_core::directory::Directory;
 use serde::{Deserialize, Serialize};
 use xi_rope::Rope;
