+++ lapce-ui/src/title.rs
@@ -8,7 +8,7 @@ use druid::{
     RenderContext, Size, Target, Widget, WidgetExt, WidgetId, WidgetPod,
     WindowConfig, WindowState,
 };
-use lapce_core::command::FocusCommand;
+use lapce_core::{command::FocusCommand, meta};
 use lapce_data::{
     command::{
         CommandKind, LapceCommand, LapceUICommand, LapceWorkbenchCommand,
@@ -19,7 +19,7 @@ use lapce_data::{
     list::ListData,
     menu::{MenuItem, MenuKind},
     palette::PaletteStatus,
-    proxy::{ProxyStatus, VERSION},
+    proxy::ProxyStatus,
 };
 
 #[cfg(not(target_os = "macos"))]
@@ -417,7 +417,7 @@ impl Title {
                 desc: Some(if data.update_in_progress && latest_version.is_some() {
                     format!("Update in progress ({}) ", latest_version.unwrap())
                 } else if latest_version.is_some()
-                    && latest_version != Some(*VERSION)
+                    && latest_version != Some(*meta::VERSION)
                 {
                     format!("Restart to update ({})", latest_version.unwrap())
                 } else {
@@ -430,7 +430,7 @@ impl Title {
                     data: None,
                 },
                 enabled: latest_version.is_some()
-                    && latest_version != Some(*VERSION)
+                    && latest_version != Some(*meta::VERSION)
                     && !data.update_in_progress,
             }),
             MenuKind::Separator,
@@ -443,7 +443,7 @@ impl Title {
                 enabled: true,
             }),
         ];
-        if latest_version.is_some() && latest_version != Some(*VERSION) {
+        if latest_version.is_some() && latest_version != Some(*meta::VERSION) {
             let text_layout = piet_text
                 .new_text_layout("1")
                 .font(data.config.ui.font_family(), 10.0)
